pyproject.toml
[project]
name = "wasfaty-pref-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.42",
    "jinja2>=3.1.6",
    "locust>=2.37.11",
    "progress>=1.6",
    "psycopg>=3.2.9",
    "psycopg-pool>=3.2.6",
    "ruff>=0.12.0",
    "jwt",
    "pyjwt>=2.10.1",
    "aiohttp>=3.13.1",
    "colorlog>=6.10.1",
    "pyyaml>=6.0.3",
    "faker>=37.12.0",
]
